requirements.txt
tensorflow>=2.12.0
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=8.3.0
